+++ src/codetext/parser/language_parser.py
@@ -4,9 +4,11 @@ from typing import List, Dict, Any, Set, Optional
 
 import tree_sitter
 
+import logging
+
 DOCSTRING_REGEX = re.compile(r"(['\"])\1\1(.*?)\1{3}", flags=re.DOTALL)
 DOCSTRING_REGEX_TOKENIZER = re.compile(r"[^\s,'\"`.():\[\]=*;>{\}+-/\\]+|\\+|\.+|\(\)|{\}|\[\]|\(+|\)+|:+|\[+|\]+|{+|\}+|=+|\*+|;+|>+|\++|-+|/+|\'|\"|`")
-
+logger = logging.getLogger()
 
 def remove_words_in_string(words, string):
     new_string = string
@@ -109,15 +111,53 @@ def traverse(node, results: List) -> None:
 
 
 def traverse_type(node, results, kind:List) -> None:
+    # logger.warn('From version 0.0.6, we move `traverse_type` to `get_node_by_kind`')
     if node.type in kind:
         results.append(node)
     if not node.children:
         return
     for n in node.children:
         traverse_type(n, results, kind)
+
+
+def get_node_by_kind(root: tree_sitter.Node, kind: List[str]) -> List:
+    """
+    Get all nodes with specific type
+    
+    Args:
+        root (tree_sitter.Node): Tree sitter root node
+        kind (List[str]): (node's) type that want to get
+    
+    Return:
+        List[tree_sitter.Node]: List of all 
+    """
+    assert type(root) == tree_sitter.Node, f"Expect `root` to be `tree_sitter.Node`, get {type(root)}"
+    assert type(kind) in [list, str], f"Expect `kind` to be `list` of string or `str`, get {type(kind)}"
+    assert all(isinstance(s, str) for s in kind) == True, f"Expect search kind to be `str`"
+
+    node_list = []
+    traverse_type(root, node_list, kind=kind)
+    return node_list
+
+
+def get_node_text(root: tree_sitter.Node) -> str:
+    """
+    Get text of a tree-sitter Node. Can be use to replace `match_from_span`.
+    
+    Args:
+        root (tree_sitter.Node): Tree sitter node to get text
         
+    Return:
+        str: text of `root`
+    """
+    assert type(root) == tree_sitter.Node, f"Expect `root` to be `tree_sitter.Node`, get {type(root)}"
+
+    text = root.text.decode()
+    return text
+
 
 def match_from_span(node, blob: str) -> str:
+    # logger.warn('From version 0.0.6, we move `match_from_span` to `get_node_text`')
     lines = blob.split('\n')
     line_start = node.start_point[0]
     line_end = node.end_point[0]
@@ -127,6 +167,41 @@ def match_from_span(node, blob: str) -> str:
         return '\n'.join([lines[line_start][char_start:]] + lines[line_start+1:line_end] + [lines[line_end][:char_end]])
     else:
         return lines[line_start][char_start:char_end]
+    
+
+def match_from_spans(nodes, blob: str) -> str:
+    """
+    Get text from multiple note
+    
+    Args:
+        nodes (List): List of `tree_sitter.Node`
+        blob (str): Full source
+    
+    Return:
+        str: combined text of list node
+    """
+    assert len(nodes) != 0, "Empty node list"
+    start_point = nodes[0]
+    end_point = nodes[0]
+    
+    for node in nodes:
+        if node.start_point[0] < start_point.start_point[0]:
+            start_point = node
+        elif node.end_point[0] > end_point.end_point[0]:
+            end_point = node
+    
+    line_start = start_point.start_point[0]
+    char_start = start_point.start_point[1]
+    line_end = end_point.end_point[0]
+    char_end = end_point.end_point[1]
+        
+    lines = blob.split('\n')
+    if line_start != line_end:
+        string = '\n'.join([lines[line_start][char_start:]] + lines[line_start+1:line_end] + [lines[line_end][:char_end]])
+    else:
+        string = lines[line_start][char_start:char_end]
+    
+    return string, start_point, end_point
 
 
 class LanguageParser(ABC):