+++ src/codetext/parser/ruby_parser.py
@@ -2,9 +2,13 @@ import re
 from typing import List, Dict, Any
 
 import tree_sitter
+import logging
 
-from .language_parser import LanguageParser, match_from_span, tokenize_code, tokenize_docstring, traverse_type
+from .language_parser import LanguageParser, get_node_text, get_node_by_kind
 # from function_parser.parsers.commentutils import get_docstring_summary
+
+
+logger = logging.getLogger(__name__)
 
 
 class RubyParser(LanguageParser):
@@ -16,14 +20,12 @@ class RubyParser(LanguageParser):
 
     @staticmethod
     def get_function_list(node):
-        res = []
-        traverse_type(node, res, ['method'])
+        res = get_node_by_kind(node, ['method'])
         return res
     
     @staticmethod
     def get_class_list(node):
-        res = []
-        traverse_type(node, res, ['class', 'module'])
+        res = get_node_by_kind(node, ['class', 'module'])
         
         # remove class keywords
         for node in res[:]:
@@ -60,10 +62,11 @@ class RubyParser(LanguageParser):
     
     @staticmethod
     def get_docstring(node, blob):
+        logger.info('From version `0.0.6` this function will update argument in the API')
         docstring_node = RubyParser.get_docstring_node(node)
         docstring = []
         for item in docstring_node:
-            doc = match_from_span(item, blob)
+            doc = get_node_text(item)
             doc_lines = doc.split('\n')
             for line in doc_lines:
                 if '=begin' in line or '=end' in line:
@@ -75,9 +78,11 @@ class RubyParser(LanguageParser):
     
     @staticmethod
     def get_function_metadata(function_node, blob) -> Dict[str, str]:
+        logger.info('From version `0.0.6` this function will update argument in the API')
         metadata = {
             'identifier': '',
             'parameters': [],
+            'return_type': None,
         }
         
         assert type(function_node) == tree_sitter.Node
@@ -85,17 +90,17 @@ class RubyParser(LanguageParser):
         
         for child in function_node.children:
             if child.type == 'identifier':
-                metadata['identifier'] = match_from_span(child, blob)
+                metadata['identifier'] = get_node_text(child)
             elif child.type in ['method_parameters', 'parameters', 'bare_parameters']:
-                params = []
-                traverse_type(child, params, ['identifier'])
+                params = get_node_by_kind(child, ['identifier'])
                 for item in params:
-                    metadata['parameters'].append(match_from_span(item, blob))
+                    metadata['parameters'].append(get_node_text(item))
 
         return metadata
     
     @staticmethod
     def get_class_metadata(class_node, blob):
+        logger.info('From version `0.0.6` this function will update argument in the API')
         metadata = {
             'identifier': '',
             'parameters': [],
@@ -105,17 +110,16 @@ class RubyParser(LanguageParser):
         
         for child in class_node.children:
             if child.type == 'constant':
-                metadata['identifier'] = match_from_span(child, blob)
+                metadata['identifier'] = get_node_text(child)
             if child.type == 'superclass':
                 for subchild in child.children:
                     if subchild.type == 'constant':
-                        metadata['parameters'].append(match_from_span(subchild, blob))
+                        metadata['parameters'].append(get_node_text(subchild))
 
         return metadata
         
 
     @staticmethod
     def get_comment_node(function_node):
-        comment_node = []
-        traverse_type(function_node, comment_node, kind='comment')
+        comment_node = get_node_by_kind(function_node, kind='comment')
         return comment_node